Changed from directly interacting with GSON to using POJOs, and made provider plugin more proactive.
Changed from interacting directly with GSON to use regular objects which can be serialized by GSON.
This provides a more convenient way of standardizing the schema, and allows for aspects to be used
more efficiently. This creates an easy way to add metadata to queries, as well as timing information.
The provider plugin interface was modified to have an option to invoke an action on object creation
rather than waiting for everything to finish. This was in response to an out-of-memory issue discovered
when importing large repositories.