Forced JGit to put the .git folder along side the repo.
For some weird reason, JGit will put the .git directory in the local directory
rather than along side the rest of the repo. This was causing issues when running
tests and generating a .git folder inside the GitProviderPlugin folder, making the
subfolder it's own git repo pointing to the most recently run. This commit forces
the .git directory to side beside the repo, fixing testing and probably making this
more thread-safe as well.