Standardizing the elasticsearch schema.
This commit standardizes the output schema. The recommended elasticsearch mapping
is included in the es_schema.json file. As a result, the JavaCodeUnit class was modified
to change the annotation map from an arbitrary object to be a Map<String, String>, changing
old behavior. In annotations, old behavior was to use a string for an annotation not naming
key, rather than using a map. New behavior is a map with key "#body", with the value being
the unkeyed parameter.