Correcting behavior when encountering unusual locations for javadoc comments.
This commit fixes behavior where javadoc comments in unusual locations
(before another javadoc comment, in places where a javadoc comment doesn't
sense like the middle of a method) would cause the output to be unusual.
Behavior now matches that of javadoc, which is to ignore all but the last
javadoc comment in situations where a javadoc comment is expected, and
ignore all others.