Added support for custom behavior when loading properties
Added support for a PluginLoader to notice specially named properties
and to run a function when special properties are requested. This allows
for things like database queries to be run when loading a plugin. Also added
a tree structure in PackageHierarchy, which allows for modeling a tree hierarchy
for package structures, such as in Java.