Adds PackageHierarchy management into plugins, fixed parse issues.
This commit adds the ability for the plugin to request a known
PackageHierarchy over a specific field. This is used to better manage
star imports in Java, as well as classes within the same package.
While this improves the ability for plugins to predict fully qualified
name correctly, it is *NOT* a complete solution. It does not fix load
order issues, among others. A way to fix this would be to index everything
twice (once to populate the package hierarchy, another to get everything
right).