Fixes an issue where generic types were prevented classes from passing through the DependencyManager.
Generic types were being added onto the class names, leading to dependencies on
'foo.Bar<T>'. Since generics aren't part of the class name, this prevented classes
from having all dependencies filled, and thus sitting in the dependency manager
forever. This also adds support for generic types, as classes using a generic type
would get stuck in the dependency manager waiting on a class with the name of their
generic to be added, which wouldn't happen.